Developing algorithm to identify disjoint components in the graph.
The algorithm is working fine, but there a bug I can't find. I think it is connencted with insertion and deletion of elements in the components map, but I can't find where the bug is. The algorithm seems fine also when deleting components in components_map using erase().
Otherwise I don't really know.

There also are some nomenclature to modify and possible new methods to implement (like the insertion of one existing component in another existing component, of which I give a first rough implementation, but not yet usable)